pyproject.toml
[project]
name = "memcoach"
version = "0.1.0"
description = "Steering-based memorability coaching for multimodal models."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "accelerate>=1.12.0",
    "colorlog>=6.10.1",
    "datasets>=4.6.0",
    "diffusers>=0.36.0",
    "fastapi>=0.128.0",
    "fire>=0.7.1",
    "flash-attn>=2.8.3",
    "huggingface-hub",
    "hydra-core>=1.3.2",
    "json-repair>=0.54.3",
    "litellm>=1.80.10",
    "ngrok>=1.7.0",
    "open-clip-torch>=3.2.0",
    "outlines>=1.2.9",
    "pandas>=3.0.0",
    "protobuf>=6.33.2",
    "python-multipart>=0.0.21",
    "sentencepiece>=0.2.1",
    "timm>=1.0.22",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers==4.57.3",
    "uvicorn[standard]>=0.40.0",
]

[tool.uv]
no-build-isolation-package = ["flash-attn"]
